adjust child transform when removing or setting the parent layer
